refactor: moved PopoverMenu into its own component …
… instead of using memo. From performance perspective that even seems to have slightly improved over the React.memo approach.
