+++ src/pages/home/sidebar/SidebarScreen/index.native.js
@@ -1,28 +1,20 @@
 import React from 'react';
-import {withOnyx} from 'react-native-onyx';
-import compose from '../../../../libs/compose';
-import withWindowDimensions from '../../../../components/withWindowDimensions';
-import withLocalize from '../../../../components/withLocalize';
-import ONYXKEYS from '../../../../ONYXKEYS';
 import {sidebarPropTypes, sidebarDefaultProps} from './sidebarPropTypes';
 import BaseSidebarScreen from './BaseSidebarScreen';
+import PopoverModal from './PopoverModal';
 
-// eslint-disable-next-line react/jsx-props-no-spreading
-const SidebarScreen = props => <BaseSidebarScreen {...props} />;
+const SidebarScreen = props => (
+    <>
+        <BaseSidebarScreen
+            // eslint-disable-next-line react/jsx-props-no-spreading
+            {...props}
+        />
+        <PopoverModal />
+    </>
+);
 
 SidebarScreen.propTypes = sidebarPropTypes;
 SidebarScreen.defaultProps = sidebarDefaultProps;
 SidebarScreen.displayName = 'SidebarScreen';
 
-export default compose(
-    withLocalize,
-    withWindowDimensions,
-    withOnyx({
-        allPolicies: {
-            key: ONYXKEYS.COLLECTION.POLICY,
-        },
-        betas: {
-            key: ONYXKEYS.BETAS,
-        },
-    }),
-)(SidebarScreen);
+export default SidebarScreen;
